requirements.txt
google-api-python-client>=2.70.0
google-auth>=2.20.0
newspaper3k>=0.2.8
sentence-transformers>=2.2.2
numpy>=1.25.0
openai>=0.27.0
jinja2>=3.1.2
schedule>=1.2.0
pydantic>=1.10.0
pytest>=7.0.0
flake8>=5.0.0